split Sidebar sub-components into files
